Algo: Basic algorithm for fitness
Can now score DNA with fitness based on a direction of movement. The generations created are tested and the best of each round are stored in a list. After the generations are finished, the best bots are created and instantiated. This is a very simple representation which is to be expanded on.