Algo: Resolved Instantiation Problem
The problem occurs behind the scenes in unity. The naming convention used by the program, searches for a block by name, and if the array is not named uniquely, regardless of if the block is deleted, it will attempt to attach to the first iteration block, which is not present in the current generation.  This was a terrible bug to try to find, but by giving the Creatures names that are unique by generation, the instantiation conflicts are resolved.