Algo: Single Creature Mutation Live!
Program is now capable of selecting a winner from each generation, then creating a generation from that winner. This process is repeated with mutation levels set in the editor. Results in the creation of a winning "walking" bot. This is considered phase 1 complete. A functioning prototype that evolves through mutation to achieve the highest fitness score.