Algo: Multi-Winners and Standard Deviations
Winners are now selected as a sqrt of the number of creatures from each generation. Also started implementing a procedure for measuring the standard deviation of the blocks in order to eliminate the blocks that seperate into multiple pieces.  This is not fully implemented at this point.  Also need to revise the WinnerList into a hard coded WinnerArray and keep a consistent format.